--disable-subtract

+++ eg/alter_for_float_number.pl
@@ -19,6 +19,7 @@ Getopt::Long::Configure ("no_ignore_case");
 GetOptions(
     'data-dir=s' => \my $data_dir,
     'with-mysql=s' => \my $mysql,
+    'disable-subtract' => \my $disable_subtract,
     "h|help" => \my $help,
 );
 
@@ -44,12 +45,24 @@ else {
 
 my $enable_float_number = 1;
 my $data = $mysql
-    ? GrowthForecast::Data::MySQL->new($mysql, $enable_float_number)
-    : GrowthForecast::Data->new($data_dir, $enable_float_number);
+    ? GrowthForecast::Data::MySQL->new($mysql, $enable_float_number, $disable_subtract)
+    : GrowthForecast::Data->new($data_dir, $enable_float_number, $disable_subtract);
 
 if ($mysql) {
     my $number_type = $data->number_type;
     my $complex_number_type = $data->complex_number_type;
+    unless ( $disable_subtract ) {
+        $data->dbh->query(<<EOF);
+ALTER TABLE prev_graphs
+    MODIFY COLUMN number   $number_type NOT NULL DEFAULT 0,
+    MODIFY COLUMN subtract $number_type
+EOF
+        $data->dbh->query(<<EOF);
+ALTER TABLE prev_short_graphs
+    MODIFY COLUMN number   $number_type NOT NULL DEFAULT 0,
+    MODIFY COLUMN subtract $number_type
+EOF
+    }
     $data->dbh->query(<<EOF);
 ALTER TABLE graphs
     MODIFY COLUMN number  $number_type NOT NULL DEFAULT 0,
@@ -59,34 +72,28 @@ ALTER TABLE graphs
     MODIFY COLUMN sllimit $number_type NOT NULL DEFAULT 0
 EOF
     $data->dbh->query(<<EOF);
-ALTER TABLE prev_graphs
-    MODIFY COLUMN number   $number_type NOT NULL DEFAULT 0,
-    MODIFY COLUMN subtract $number_type
-EOF
-    $data->dbh->query(<<EOF);
-ALTER TABLE prev_short_graphs
-    MODIFY COLUMN number   $number_type NOT NULL DEFAULT 0,
-    MODIFY COLUMN subtract $number_type
-EOF
-    $data->dbh->query(<<EOF);
 ALTER TABLE complex_graphs
     MODIFY COLUMN number   $complex_number_type UNSIGNED NOT NULL DEFAULT 0
 EOF
 }
 else {
     # NOTE: sqlite does not support `ALTER TABLE MODIFY COLUMN`.
+    unless ( $disable_subtract ) {
+        $data->dbh->query('ALTER TABLE prev_graphs RENAME TO tmp_prev_graphs');
+        $data->dbh->query('ALTER TABLE prev_short_graphs RENAME TO tmp_prev_short_graphs');
+    }
     $data->dbh->query('ALTER TABLE graphs RENAME TO tmp_graphs');
-    $data->dbh->query('ALTER TABLE prev_graphs RENAME TO tmp_prev_graphs');
-    $data->dbh->query('ALTER TABLE prev_short_graphs RENAME TO tmp_prev_short_graphs');
     $data->dbh->query('ALTER TABLE complex_graphs RENAME TO tmp_complex_graphs');
     undef($data->{dbh}); # To re-create tables by calling on_connect callback
+    unless ( $disable_subtract ) {
+        $data->dbh->query('INSERT INTO prev_graphs SELECT * from tmp_prev_graphs');
+        $data->dbh->query('INSERT INTO prev_short_graphs SELECT * from tmp_prev_short_graphs');
+        $data->dbh->query('DROP TABLE tmp_prev_graphs');
+        $data->dbh->query('DROP TABLE tmp_prev_short_graphs');
+    }
     $data->dbh->query('INSERT INTO graphs SELECT * from tmp_graphs');
-    $data->dbh->query('INSERT INTO prev_graphs SELECT * from tmp_prev_graphs');
-    $data->dbh->query('INSERT INTO prev_short_graphs SELECT * from tmp_prev_short_graphs');
-    $data->dbh->query('INSERT INTO complex_graphs SELECT * from tmp_complex_graphs');
     $data->dbh->query('DROP TABLE tmp_graphs');
-    $data->dbh->query('DROP TABLE tmp_prev_graphs');
-    $data->dbh->query('DROP TABLE tmp_prev_short_graphs');
+    $data->dbh->query('INSERT INTO complex_graphs SELECT * from tmp_complex_graphs');
     $data->dbh->query('DROP TABLE tmp_complex_graphs');
 }
 
@@ -123,6 +130,10 @@ $ alter_for_float_number.pl
 
  $ MYSQL_USER=www MYSQL_PASSWORD=foobar alter_for_float_number.pl \\
      --with-mysql dbi:mysql:growthforecast;hostname=localhost
+
+=item --disable-subtract
+
+ Specify if your GrowthForecast is using --disable-subtract option.
 
 =item -h --help
